# internal/queryapi/playground/openapi.yaml
openapi: 3.1.0
info:
  title: Logagg API
  version: 1.0.0
  description: Multi-tenant log ingestion and query API.
servers:
  - url: http://localhost:8080
    description: Ingest API
  - url: http://localhost:8081
    description: Query API
security:
  - ApiKey: []
paths:
  /v1/logs:
    post:
      summary: Ingest a batch of logs
      operationId: ingestLogs
      servers:
        - url: http://localhost:8080
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestBatch"
      responses:
        "202":
          description: Batch durably queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/Unavailable" }
    get:
      summary: Query raw logs
      operationId: queryLogs
      servers:
        - url: http://localhost:8081
      parameters:
        - $ref: "#/components/parameters/Start"
        - $ref: "#/components/parameters/End"
        - { name: service, in: query, schema: { type: string } }
        - { name: level, in: query, schema: { type: string } }
        - { name: trace_id, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 1000, default: 100 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, maximum: 10000, default: 0 } }
        - { name: stream, in: query, schema: { type: boolean, default: false } }
      responses:
        "200":
          description: Tenant-scoped query results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResponse"
            application/x-ndjson:
              schema:
                type: string
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /v1/analytics:
    get:
      summary: Run an aggregate query
      servers:
        - url: http://localhost:8081
      parameters:
        - $ref: "#/components/parameters/Start"
        - $ref: "#/components/parameters/End"
        - { name: aggregation, in: query, required: true, schema: { type: string, enum: [count, rate, percentile] } }
        - { name: bucket, in: query, schema: { type: string, enum: [minute, hour, day] } }
        - { name: group_by, in: query, schema: { type: string } }
      responses:
        "200": { description: Aggregate results }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/query:
    post:
      summary: Query using the JSON DSL
      servers:
        - url: http://localhost:8081
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200": { description: Raw or aggregate query results }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/status:
    get:
      security: []
      summary: Get query API status
      responses:
        "200": { description: Service status }
  /healthz:
    get:
      security: []
      summary: Liveness probe
      responses:
        "200": { description: Service is alive }
  /readyz:
    get:
      security: []
      summary: Dependency readiness probe
      responses:
        "200": { description: Dependencies are ready }
        "503": { description: One or more dependencies are unavailable }
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    Start:
      name: start
      in: query
      required: true
      schema: { type: string, format: date-time }
    End:
      name: end
      in: query
      required: true
      schema: { type: string, format: date-time }
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: API key is not authorized for the service or environment
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RateLimited:
      description: Rate limit or queue backpressure
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unavailable:
      description: A required dependency is unavailable
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    IngestBatch:
      type: object
      required: [schema_version, service, env, source, logs]
      properties:
        schema_version: { type: string, const: logs.ingest.v1 }
        service: { type: string }
        env: { type: string }
        source: { type: string }
        logs:
          type: array
          minItems: 1
          maxItems: 1000
          items: { $ref: "#/components/schemas/IngestLog" }
    IngestLog:
      type: object
      required: [timestamp, level, message]
      properties:
        timestamp: { type: string, format: date-time }
        level: { type: string }
        message: { type: string }
        fields: { type: object, additionalProperties: true }
    IngestResponse:
      type: object
      required: [request_id, accepted, status]
      properties:
        request_id: { type: string }
        accepted: { type: integer }
        status: { type: string, const: queued }
    QueryResponse:
      type: object
      required: [logs, count, partial]
      properties:
        logs:
          type: array
          items: { $ref: "#/components/schemas/LogRecord" }
        count: { type: integer }
        next_offset: { type: integer }
        partial: { type: boolean }
        unavailable_shards:
          type: array
          items: { type: string }
    LogRecord:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        tenant_id: { type: integer, format: uint64 }
        service: { type: string }
        environment: { type: string }
        source: { type: string }
        host: { type: string }
        level: { type: string }
        trace_id: { type: string }
        fingerprint: { type: string }
        message: { type: string }
        fields: { type: object, additionalProperties: true }
        ingest_id: { type: string }
        raw_size_bytes: { type: integer }
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }